Add new partial route structure
